requirements.txt
psutil
celery
pyyaml